+++ src/service/Produtos.js
@@ -1,5 +1,7 @@
 import { useEffect, useState } from "react";
 import api from "./api";
+import { exibirAlertaSucesso, exibirAlertaErro, exibirAlertaConfirmacao } from './alertas';
+
 
 function Produtos() {
     const [produtos, setProdutos] = useState([]);
@@ -10,7 +12,8 @@ function Produtos() {
             const response = await api.get("/produtos");
             setProdutos(response.data);
         } catch (error) {
-            console.error("Erro ao buscar:", error);
+            exibirAlertaErro("Erro ao buscar produtos.", );
+            throw error
         } finally {
             setLoading(false);
         }
@@ -19,50 +22,48 @@ function Produtos() {
     const adicionarProduto = async (dadosDoFormulario) => {
         try {
             const response = await api.post("/produtos", dadosDoFormulario);
-
             setProdutos(prevClientes => [...prevClientes, response.data]);
-
+            exibirAlertaSucesso("Produto adicionado com sucesso!");
             return response.data;
         } catch (error) {
-            console.error("Erro na requisição POST:", error);
+            exibirAlertaErro("Erro ao adicionar produto.");
             throw error;
         }
     };
 
     const excluirProduto = async (id) => {
-        try {
-            await api.delete(`/produtos/${id}`);
-
-            setProdutos(prev => prev.filter(produto => {
-                const atualId = produto.id_peca || produto.idPeca;
-                return Number(atualId) !== Number(id);
-            }));
-
-            console.log("Produto removido da tela com sucesso");
-        } catch (error) {
-            console.error("Erro ao excluir cliente:", error);
-            alert("Erro ao excluir no servidor.");
+        const confirmacao = await exibirAlertaConfirmacao("Deseja realmente excluir este produto?");
+        
+        if (confirmacao.isConfirmed) {
+            try {
+                await api.delete(`/produtos/${id}`);
+                setProdutos(prev => prev.filter(produto => {
+                    const atualId = produto.id_peca || produto.idPeca;
+                    return Number(atualId) !== Number(id);
+                }));
+                exibirAlertaSucesso("Produto excluído com sucesso!");
+            } catch (error) {
+                exibirAlertaErro("Erro ao excluir produto.");
+                throw error
+            }
         }
     };
 
-const atualizarProduto = async (id, dadosAtualizados) => {
-    try {
-        const response = await api.put(`/produtos/${id}`, dadosAtualizados);
-        
-        const produtoAtualizado = response.data;
-
-        setProdutos(prev => prev.map(p => {
-            const isTarget = p.id_peca === id || p.idPeca === id || p.id === id;
-            
-            return isTarget ? produtoAtualizado : p;
-        }));
-
-        return produtoAtualizado;
-    } catch (error) {
-        console.error("Erro ao atualizar: ", error);
-        throw error;
-    }
-};
+    const atualizarProduto = async (id, dadosAtualizados) => {
+        try {
+            const response = await api.put(`/produtos/${id}`, dadosAtualizados);
+            const produtoAtualizado = response.data;
+            setProdutos(prev => prev.map(p => {
+                const isTarget = p.id_peca === id || p.idPeca === id || p.id === id;
+                return isTarget ? produtoAtualizado : p;
+            }));
+            exibirAlertaSucesso("Produto atualizado com sucesso!");
+            return produtoAtualizado;
+        } catch (error) {
+            exibirAlertaErro("Erro ao atualizar produto.");
+            throw error;
+        }
+    };
 
     useEffect(() => {
         listarProdutos();
@@ -71,4 +72,4 @@ const atualizarProduto = async (id, dadosAtualizados) => {
     return { produtos, loading, adicionarProduto, excluirProduto, atualizarProduto };
 }
 
-export default Produtos
+export default Produtos;
